requirements.txt
langchain
langgraph
langgraph-swarm
langchain-google-genai
google-generativeai
langchain_community
faiss-cpu
tavily-python
google-cloud-speech
sounddevice
portaudio19-dev
scipy
pdfminer.six
python-dotenv
langchain-openai
numpy
pandas
openai
pytesseract
openpyxl